+++ src/test/scala/io/archivesunleashed/ArchiveRecordTest.scala
@@ -16,8 +16,6 @@
  */
 package io.archivesunleashed
 
-import matchbox._
-
 import com.google.common.io.Resources
 import org.apache.spark.{SparkConf, SparkContext}
 import org.junit.runner.RunWith
@@ -41,8 +39,8 @@ class ArchiveRecordTest extends FunSuite with BeforeAndAfter {
   }
 
   test("count records") {
-    assert(RecordLoader.loadArchives(arcPath, sc, keepValidPages = false).count == 300L)
-    assert(RecordLoader.loadArchives(warcPath, sc, keepValidPages = false).count == 299L)
+    assert(RecordLoader.loadArchives(arcPath, sc).count == 300L)
+    assert(RecordLoader.loadArchives(warcPath, sc).count == 299L)
   }
 
   after {
