+++ src/test/scala/io/archivesunleashed/matchbox/WriteGEXFTest.scala
@@ -19,13 +19,11 @@ package io.archivesunleashed.matchbox
 
 import org.apache.spark.{ SparkConf, SparkContext }
 import org.junit.runner.RunWith
-import org.scalatest.FunSuite
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{ BeforeAndAfter, FunSuite }
 import java.io.File
 import java.nio.file.{Paths, Files}
-import io.archivesunleashed.matchbox._
-import io.archivesunleashed.rdd.RecordRDD._
+import io.archivesunleashed.Transformations._
 import scala.io.Source
 
 @RunWith(classOf[JUnitRunner])
